Made StreamFeed into it's own component that shouuuld be reusable.
